+++ models/inception_v4/model.py
@@ -1,333 +1,23 @@
-""" Pytorch Inception-V4 implementation
-Sourced from https://github.com/Cadene/pretrained-models.pytorch (MIT License)
-"""
-
-from __future__ import print_function, division, absolute_import
-import os
-import sys
-
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
+import json
 import torch.utils.model_zoo as model_zoo
+from torch import nn
 
-from torchsummary import summary
-
-
-pretrained_settings = {
-    'inceptionv4': {
-        'imagenet': {
-            'url': 'http://data.lip6.fr/cadene/pretrainedmodels/inceptionv4-8e4777a0.pth',
-            'input_space': 'RGB',
-            'input_size': [3, 299, 299],
-            'input_range': [0, 1],
-            'mean': [0.5, 0.5, 0.5],
-            'std': [0.5, 0.5, 0.5],
-            'num_classes': 1000
-        },
-        'imagenet+background': {
-            'url': 'http://data.lip6.fr/cadene/pretrainedmodels/inceptionv4-8e4777a0.pth',
-            'input_space': 'RGB',
-            'input_size': [3, 299, 299],
-            'input_range': [0, 1],
-            'mean': [0.5, 0.5, 0.5],
-            'std': [0.5, 0.5, 0.5],
-            'num_classes': 1001
-        }
-    }
-}
-
-
-class BasicConv2d(nn.Module):
-
-    def __init__(self, in_planes, out_planes, kernel_size, stride, padding=0) -> None:
-        super(BasicConv2d, self).__init__()
-        self.conv = nn.Conv2d(in_planes, out_planes,
-                              kernel_size=kernel_size, stride=stride,
-                              padding=padding, bias=False) # verify bias false
-        self.bn = nn.BatchNorm2d(out_planes,
-                                 eps=0.001, # value found in tensorflow
-                                 momentum=0.1, # default pytorch value
-                                 affine=True)
-        self.relu = nn.ReLU(inplace=True)
-
-    def forward(self, x) -> torch.Tensor:
-        x = self.conv(x)
-        x = self.bn(x)
-        x = self.relu(x)
-        return x
-
-
-class Mixed_3a(nn.Module):
-
-    def __init__(self) -> None:
-        super(Mixed_3a, self).__init__()
-        self.maxpool = nn.MaxPool2d(3, stride=2)
-        self.conv = BasicConv2d(64, 96, kernel_size=3, stride=2)
-
-    def forward(self, x) -> torch.Tensor:
-        x0 = self.maxpool(x)
-        x1 = self.conv(x)
-        out = torch.cat((x0, x1), 1)
-        return out
-
-
-class Mixed_4a(nn.Module):
-
-    def __init__(self) -> None:
-        super(Mixed_4a, self).__init__()
-
-        self.branch0 = nn.Sequential(
-            BasicConv2d(160, 64, kernel_size=1, stride=1),
-            BasicConv2d(64, 96, kernel_size=3, stride=1)
-        )
-
-        self.branch1 = nn.Sequential(
-            BasicConv2d(160, 64, kernel_size=1, stride=1),
-            BasicConv2d(64, 64, kernel_size=(1,7), stride=1, padding=(0,3)),
-            BasicConv2d(64, 64, kernel_size=(7,1), stride=1, padding=(3,0)),
-            BasicConv2d(64, 96, kernel_size=(3,3), stride=1)
-        )
-
-    def forward(self, x) -> torch.Tensor:
-        x0 = self.branch0(x)
-        x1 = self.branch1(x)
-        out = torch.cat((x0, x1), 1)
-        return out
-
-
-class Mixed_5a(nn.Module):
-
-    def __init__(self) -> None:
-        super(Mixed_5a, self).__init__()
-        self.conv = BasicConv2d(192, 192, kernel_size=3, stride=2)
-        self.maxpool = nn.MaxPool2d(3, stride=2)
-
-    def forward(self, x) -> torch.Tensor:
-        x0 = self.conv(x)
-        x1 = self.maxpool(x)
-        out = torch.cat((x0, x1), 1)
-        return out
-
-
-class Inception_A(nn.Module):
-
-    def __init__(self) -> None:
-        super(Inception_A, self).__init__()
-        self.branch0 = BasicConv2d(384, 96, kernel_size=1, stride=1)
-
-        self.branch1 = nn.Sequential(
-            BasicConv2d(384, 64, kernel_size=1, stride=1),
-            BasicConv2d(64, 96, kernel_size=3, stride=1, padding=1)
-        )
-
-        self.branch2 = nn.Sequential(
-            BasicConv2d(384, 64, kernel_size=1, stride=1),
-            BasicConv2d(64, 96, kernel_size=3, stride=1, padding=1),
-            BasicConv2d(96, 96, kernel_size=3, stride=1, padding=1)
-        )
-
-        self.branch3 = nn.Sequential(
-            nn.AvgPool2d(3, stride=1, padding=1, count_include_pad=False),
-            BasicConv2d(384, 96, kernel_size=1, stride=1)
-        )
-
-    def forward(self, x) -> torch.Tensor:
-        x0 = self.branch0(x)
-        x1 = self.branch1(x)
-        x2 = self.branch2(x)
-        x3 = self.branch3(x)
-        out = torch.cat((x0, x1, x2, x3), 1)
-        return out
-
-
-class Reduction_A(nn.Module):
-
-    def __init__(self) -> None:
-        super(Reduction_A, self).__init__()
-        self.branch0 = BasicConv2d(384, 384, kernel_size=3, stride=2)
-
-        self.branch1 = nn.Sequential(
-            BasicConv2d(384, 192, kernel_size=1, stride=1),
-            BasicConv2d(192, 224, kernel_size=3, stride=1, padding=1),
-            BasicConv2d(224, 256, kernel_size=3, stride=2)
-        )
-
-        self.branch2 = nn.MaxPool2d(3, stride=2)
-
-    def forward(self, x) -> torch.Tensor:
-        x0 = self.branch0(x)
-        x1 = self.branch1(x)
-        x2 = self.branch2(x)
-        out = torch.cat((x0, x1, x2), 1)
-        return out
-
-
-class Inception_B(nn.Module):
-
-    def __init__(self) -> None:
-        super(Inception_B, self).__init__()
-        self.branch0 = BasicConv2d(1024, 384, kernel_size=1, stride=1)
-
-        self.branch1 = nn.Sequential(
-            BasicConv2d(1024, 192, kernel_size=1, stride=1),
-            BasicConv2d(192, 224, kernel_size=(1,7), stride=1, padding=(0,3)),
-            BasicConv2d(224, 256, kernel_size=(7,1), stride=1, padding=(3,0))
-        )
-
-        self.branch2 = nn.Sequential(
-            BasicConv2d(1024, 192, kernel_size=1, stride=1),
-            BasicConv2d(192, 192, kernel_size=(7,1), stride=1, padding=(3,0)),
-            BasicConv2d(192, 224, kernel_size=(1,7), stride=1, padding=(0,3)),
-            BasicConv2d(224, 224, kernel_size=(7,1), stride=1, padding=(3,0)),
-            BasicConv2d(224, 256, kernel_size=(1,7), stride=1, padding=(0,3))
-        )
-
-        self.branch3 = nn.Sequential(
-            nn.AvgPool2d(3, stride=1, padding=1, count_include_pad=False),
-            BasicConv2d(1024, 128, kernel_size=1, stride=1)
-        )
-
-    def forward(self, x) -> torch.Tensor:
-        x0 = self.branch0(x)
-        x1 = self.branch1(x)
-        x2 = self.branch2(x)
-        x3 = self.branch3(x)
-        out = torch.cat((x0, x1, x2, x3), 1)
-        return out
-
-
-class Reduction_B(nn.Module):
-
-    def __init__(self) -> None:
-        super(Reduction_B, self).__init__()
-
-        self.branch0 = nn.Sequential(
-            BasicConv2d(1024, 192, kernel_size=1, stride=1),
-            BasicConv2d(192, 192, kernel_size=3, stride=2)
-        )
-
-        self.branch1 = nn.Sequential(
-            BasicConv2d(1024, 256, kernel_size=1, stride=1),
-            BasicConv2d(256, 256, kernel_size=(1,7), stride=1, padding=(0,3)),
-            BasicConv2d(256, 320, kernel_size=(7,1), stride=1, padding=(3,0)),
-            BasicConv2d(320, 320, kernel_size=3, stride=2)
-        )
-
-        self.branch2 = nn.MaxPool2d(3, stride=2)
-
-    def forward(self, x) -> torch.Tensor:
-        x0 = self.branch0(x)
-        x1 = self.branch1(x)
-        x2 = self.branch2(x)
-        out = torch.cat((x0, x1, x2), 1)
-        return out
-
-
-class Inception_C(nn.Module):
-
-    def __init__(self) -> None:
-        super(Inception_C, self).__init__()
-
-        self.branch0 = BasicConv2d(1536, 256, kernel_size=1, stride=1)
-
-        self.branch1_0 = BasicConv2d(1536, 384, kernel_size=1, stride=1)
-        self.branch1_1a = BasicConv2d(384, 256, kernel_size=(1,3), stride=1, padding=(0,1))
-        self.branch1_1b = BasicConv2d(384, 256, kernel_size=(3,1), stride=1, padding=(1,0))
-
-        self.branch2_0 = BasicConv2d(1536, 384, kernel_size=1, stride=1)
-        self.branch2_1 = BasicConv2d(384, 448, kernel_size=(3,1), stride=1, padding=(1,0))
-        self.branch2_2 = BasicConv2d(448, 512, kernel_size=(1,3), stride=1, padding=(0,1))
-        self.branch2_3a = BasicConv2d(512, 256, kernel_size=(1,3), stride=1, padding=(0,1))
-        self.branch2_3b = BasicConv2d(512, 256, kernel_size=(3,1), stride=1, padding=(1,0))
-
-        self.branch3 = nn.Sequential(
-            nn.AvgPool2d(3, stride=1, padding=1, count_include_pad=False),
-            BasicConv2d(1536, 256, kernel_size=1, stride=1)
-        )
-
-    def forward(self, x) -> torch.Tensor:
-        x0 = self.branch0(x)
-
-        x1_0 = self.branch1_0(x)
-        x1_1a = self.branch1_1a(x1_0)
-        x1_1b = self.branch1_1b(x1_0)
-        x1 = torch.cat((x1_1a, x1_1b), 1)
-
-        x2_0 = self.branch2_0(x)
-        x2_1 = self.branch2_1(x2_0)
-        x2_2 = self.branch2_2(x2_1)
-        x2_3a = self.branch2_3a(x2_2)
-        x2_3b = self.branch2_3b(x2_2)
-        x2 = torch.cat((x2_3a, x2_3b), 1)
-
-        x3 = self.branch3(x)
-
-        out = torch.cat((x0, x1, x2, x3), 1)
-        return out
-
-
-class Inception(nn.Module):
-    """Inception v4 model architecture from
-    `"Inception-v4, Inception-ResNet and the Impact of Residual Connections on Learning" <https://arxiv.org/abs/1602.07261>`_.
-    .. note::
-        **Important**: In contrast to the other models the inception_v4 expects tensors with a size of
-        N x 3 x 299 x 299, so ensure your images are sized accordingly.
-    """
-    def __init__(self, num_classes=1001):
-        super(Inception, self).__init__()
-        # Special attributs
-        self.input_space = None
-        self.input_size = (299, 299, 3)
-        self.mean = None
-        self.std = None
-        # Modules
-        self.features = nn.Sequential(
-            BasicConv2d(3, 32, kernel_size=3, stride=2),
-            BasicConv2d(32, 32, kernel_size=3, stride=1),
-            BasicConv2d(32, 64, kernel_size=3, stride=1, padding=1),
-            Mixed_3a(),
-            Mixed_4a(),
-            Mixed_5a(),
-            Inception_A(),
-            Inception_A(),
-            Inception_A(),
-            Inception_A(),
-            Reduction_A(), # Mixed_6a
-            Inception_B(),
-            Inception_B(),
-            Inception_B(),
-            Inception_B(),
-            Inception_B(),
-            Inception_B(),
-            Inception_B(),
-            Reduction_B(), # Mixed_7a
-            Inception_C(),
-            Inception_C(),
-            Inception_C()
-        )
-        self.last_linear = nn.Linear(1536, num_classes)
-
-    def logits(self, features):
-        #Allows image of any size to be processed
-        adaptiveAvgPoolWidth = features.shape[2]
-        x = F.avg_pool2d(features, kernel_size=adaptiveAvgPoolWidth)
-        x = x.view(x.size(0), -1)
-        x = self.last_linear(x)
-        return x
-
-    def forward(self, input):
-        x = self.features(input)
-        x = self.logits(x)
-        return x
+import inception
 
 
 class InceptionV4:
-    """Initialize Inceptionv4 model.
+    """Loads Inceptionv4 model.
     """
-    def __init__(self, num_classes=1000, pretrained='imagenet') -> None:
+    def __init__(self, num_classes=1000, config='config.json', pretrained='imagenet') -> None:
+        """Initialize inception v4 model
+        """
         self._model = None
-        self.load_model(num_classes, pretrained)
+        
+        # load config
+        with open(config) as f:
+            self.__config = json.load(f)
+
+        self.__load_model(num_classes, pretrained)
         
         # Set model to eval mode
         self._model.cuda()
@@ -347,7 +37,7 @@ class InceptionV4:
         """
         print(summary(self._model, input_size=(3, 299, 299)))
 
-    def load_model(self, num_classes, pretrained) -> None:
+    def __load_model(self, num_classes: int, pretrained: str) -> None:
         """Creates the inceptionv4 model and loads state dict.
 
         Args:
@@ -355,13 +45,13 @@ class InceptionV4:
             pretrained (bool): If True, returns a model pre-trained on ImageNet. 
         """
         if pretrained:
-            settings = pretrained_settings['inceptionv4'][pretrained]
+            settings = self.__config['inceptionv4'][pretrained]
 
             if num_classes != settings['num_classes']:
                 raise ValueError(f'num_classes should be {settings["num_classes"]}, but is {num_classes}')
 
             # both 'imagenet'&'imagenet+background' are loaded from same parameters
-            self._model = Inception(num_classes=1001)
+            self._model = inception.Inception_v4(num_classes=1001)
             self._model.load_state_dict(model_zoo.load_url(settings['url']))
 
             if pretrained == 'imagenet':
@@ -376,7 +66,7 @@ class InceptionV4:
             self._model.mean = settings['mean']
             self._model.std = settings['std']
         else:
-            self._model = Inception(num_classes=num_classes)
+            self._model = inception.Inception_v4(num_classes=num_classes)
 
 
 if __name__ == '__main__':
